+++ apps/rest-api-server/src/openapi-trpc.ts
@@ -7,6 +7,7 @@ import {
 } from "trpc-openapi";
 
 import { createAppRouter, createTRPCContext } from "@blobscan/api";
+import { BlobStorageManager } from "@blobscan/blob-storage-manager";
 import { env } from "@blobscan/env";
 import { logger } from "@blobscan/logger";
 
@@ -15,6 +16,7 @@ import { prisma } from "./clients/prisma";
 import { redis } from "./clients/redis";
 import { getBlobPropagator } from "./services/blob-propagator";
 import { getIpfsStorage } from "./services/ipfs-storage";
+import { createBlobStorages } from "./services/blob-storages";
 
 export async function setUpOpenApiTRPC(app: Express): Promise<void> {
   const appRouter = createAppRouter({
@@ -36,9 +38,14 @@ export async function setUpOpenApiTRPC(app: Express): Promise<void> {
   });
   const blobPropagator = await getBlobPropagator();
   const ipfsStorage = await getIpfsStorage();
+  const blobStorages = await createBlobStorages();
+  const blobStorageManager = blobStorages.length
+    ? new BlobStorageManager(blobStorages)
+    : undefined;
   const createContext = createTRPCContext({
     blobPropagator,
     ipfsStorage,
+    blobStorageManager,
     chain: chain,
     rollupRegistry,
     enableTracing: env.TRACES_ENABLED,