+++ packages/api/src/routers/tx/getAll.ts
@@ -13,6 +13,7 @@ import {
   withPaginationSchema,
   withPagination,
 } from "../../middlewares/withPagination";
+import { withBlobSignedUrls } from "../../middlewares/withBlobSignedUrls";
 import { publicProcedure } from "../../procedures";
 import { normalize } from "../../utils";
 import { countTxs } from "./getCount";
@@ -46,6 +47,7 @@ export const getAll = publicProcedure
   .input(withPaginationSchema)
   .use(withPagination)
   .output(outputSchema)
+  .use(withBlobSignedUrls)
   .query(
     async ({
       ctx: { prisma, expands, filters, pagination, count, rollupRegistry },